Port no-unnecessary-generics
