+++ packages/eslint-plugin/test/fixtures/types/no-unnecessary-generics/bad.d.ts
@@ -1,0 +1,2 @@
+type Fn = <T>() => T;
+type Ctr = new<T>() => T;
